+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix `js-obj-reader` default `js-value-transform`
 - Use `:dispatch-key` on `map-reader`
 - GraphQL mutation: when given empty subquery, try to infer id
+- GraphQL supports ident transform
 
 ## [1.0.0-beta6]
 - `p/join-seq` always returns a vector