+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Use `:dispatch-key` on `map-reader`
 - GraphQL mutation: when given empty subquery, try to infer id
 - GraphQL supports ident transform
+- Merge queries support
 
 ## [1.0.0-beta6]
 - `p/join-seq` always returns a vector