+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - GraphQL mutation: when given empty subquery, try to infer id
 - GraphQL supports ident transform
 - Merge queries support
+- Fulcro network helper - batch network
 
 ## [1.0.0-beta6]
 - `p/join-seq` always returns a vector