+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - GraphQL supports ident transform
 - Merge queries support
 - Fulcro network helper - batch network
+- Add elide-ast-nodes helper
 
 ## [1.0.0-beta6]
 - `p/join-seq` always returns a vector