+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 
 ## [1.0.0-beta7-development]
-- Async parser implementation
+- Add async parser helpers
 - Fix `js-obj-reader` default `js-value-transform`
 - Use `:dispatch-key` on `map-reader`
 - GraphQL mutation: when given empty subquery, try to infer id