+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Add Fulcro network helper - GraphQL network
 - Add Fulcro network helper - Batch network
 - Add elide-ast-nodes helper
+- Async placeholder nodes
+- Profile helpers (Flame Graph generation)
 
 ## [1.0.0-beta6]
 - `p/join-seq` always returns a vector